refactor: move build & publish to common util file
