+++ docs/developers/misc/build-piece.mdx
@@ -7,7 +7,7 @@ You can use the CLI to build custom pieces for the platform. This process compil
 
 ### How It Works
 
-The CLI scans the `packages/pieces/custom/` directory for the specified piece. It checks the **name** in the `package.json` file. If the piece is found, it builds and packages it into a `.tgz` archive.
+The CLI scans the `packages/pieces/` directory for the specified piece. It checks the **name** in the `package.json` file. If the piece is found, it builds and packages it into a `.tgz` archive.
 
 ### Usage
 
@@ -17,12 +17,5 @@ To build a piece, follow these steps:
 2. Run the following command, replacing `PIECE_NAME` with the name of the piece you want to build:
 
 ```bash
-npm run build-piece -- --name PIECE_NAME
+npm run pieces -- build --name=google-sheets
 ```
-
-### Developer Workflow
-
-1. Developers create and modify the pieces offline.
-2. Increment the piece version in their corresponding `package.json`. For more information, refer to the [piece versioning](../../developers/piece-reference/piece-versioning) documentation.
-3. Open a pull request towards the main branch.
-4. Once the pull request is merged to the main branch, manually run the CLI or use a GitHub/GitLab Action to trigger the building process.
