+++ docs/developers/sharing-pieces/private.mdx
@@ -11,20 +11,19 @@ This guide assumes you have already created a piece and created a private fork o
 Friendly Tip: There is a CLI to easily upload it to your platform. Please check out [Publish Custom Pieces](../misc/publish-piece).
 </Tip>
 
+
+
 <Steps>
     <Step title="Build Piece">
         Build the piece using the following command. Make sure to replace `${name}` with your piece name.
 
         ```bash
-        npx nx build pieces-${name}
+        npm run pieces -- build --name=${name}
         ```
-    </Step>
-    <Step title="Pack the Piece">
-        Then pack your pieces as an npm package. Make sure to replace `${name}` with your piece name.
 
-        ```bash
-        cd dist/packages/pieces/${name} && npm pack
-        ```
+        <Info>
+        More information about building pieces can be found [here](../misc/build-piece).
+        </Info>
     </Step>
     <Step title="Upload Tarball">
         Upload the generated tarball inside `dist/packages/pieces/${name}`from Activepieces Platform Admin -> Pieces
