+++ packages/cli/src/lib/commands/build-piece.ts
@@ -1,24 +1,15 @@
 import { Command } from "commander";
-import { findAllPieces } from "../utils/piece-utils";
-import { readProjectJson } from "../utils/files";
-import path, { join } from "path";
-import { exec } from "../utils/exec";
+import { buildPiece, findAllPieces } from "../utils/piece-utils";
 import chalk from "chalk";
 
 async function buildPieces(pieceName: string) {
-    const piecesFolder = await findAllPieces(join(process.cwd(), 'packages', 'pieces', 'custom', pieceName));
-
+    const piecesFolder = await findAllPieces();
     if (piecesFolder.length === 0) {
         console.error(chalk.red(`Piece '${pieceName}' not found.`));
         process.exit(1);
     }
-
-    const projectJson = await readProjectJson(piecesFolder[0]);
-
-    await exec(`npx nx build ${projectJson.name} --skip-cache && cd ${path.resolve('dist/packages' + piecesFolder[0].split('/packages')[1])} && npm pack`);
-
-    const compiledPath = path.resolve('dist/packages' + piecesFolder[0].split('/packages')[1]);
-    console.info(chalk.green(`Piece '${projectJson.name}' built and packed successfully at ${compiledPath}.`));
+    const { outputFolder } = await buildPiece(piecesFolder[0]);
+    console.info(chalk.green(`Piece '${pieceName}' built and packed successfully at ${outputFolder}.`));
 }
 
 export const buildPieceCommand = new Command('build')
